refactor(host): pull config trait to top level
