+++ crates/host/src/lib.rs
@@ -8,6 +8,7 @@
 //!
 //! ## Key Modules
 //!
+//! - **[`config`]**: Implements the [`crate::config::ConfigManager`] trait for managing a configuration store that can be watched for updates. This is a supertrait of [`crate::store::StoreManager`] and is implemented by [`crate::store::DefaultStore`].
 //! - **[`event`]**: Provides the [`crate::event::EventPublisher`] trait for receiving and publishing events from the host.
 //! - **[`metrics`]**: Implements OpenTelemetry metrics for wasmCloud, primarily using [`wasmcloud_tracing`].
 //! - **[`nats`]**: Contains the NATS-based implementations for the wasmCloud host extension traits.
@@ -25,6 +26,8 @@
 //!
 //! For example, you can implement custom policies, secrets management, or registry configurations to tailor the host to your specific needs.
 //!
+//! The `wasmcloud` crate and binary uses the implementations in [crate::nats] to provide a NATS-based host runtime. This allows you to run wasmCloud components and plugins in a distributed environment, leveraging NATS for messaging and communication.
+//!
 //! ## Getting Started
 //!
 //! To get started with wasmCloud, refer to [`wasmbus`] for the core host functionality. From there, you can explore the other modules to add extensions and integrations as needed.
@@ -33,6 +36,11 @@
 
 #![warn(missing_docs)]
 #![forbid(clippy::unwrap_used)]
+
+/// [crate::config::ConfigManager] trait for managing a config store which can be watched to receive
+/// updates to the config. This is a supertrait of [crate::store::StoreManager] and is implemented
+/// by [crate::store::DefaultStore].
+pub mod config;
 
 /// [crate::event::EventPublisher] trait for receiving and publishing events from the host
 pub mod event;
